chore: add mock upload functionality
